refactor(core): convert apps.management.reset_device to layouts
