+++ core/src/apps/management/reset_device/__init__.py
@@ -88,7 +88,7 @@ async def reset_device(ctx: wire.Context, msg: ResetDevice) -> Success:
         secret,  # for SLIP-39, this is the EMS
         msg.backup_type,
         needs_backup=not perform_backup,
-        no_backup=msg.no_backup,
+        no_backup=bool(msg.no_backup),
     )
 
     # if we backed up the wallet, show success message
@@ -109,12 +109,17 @@ async def backup_slip39_basic(
     await layout.slip39_show_checklist(ctx, 1, BackupType.Slip39_Basic)
     threshold = await layout.slip39_prompt_threshold(ctx, shares_count)
 
+    identifier = storage.device.get_slip39_identifier()
+    iteration_exponent = storage.device.get_slip39_iteration_exponent()
+    if identifier is None or iteration_exponent is None:
+        raise ValueError
+
     # generate the mnemonics
     mnemonics = slip39.split_ems(
         1,  # Single Group threshold
         [(threshold, shares_count)],  # Single Group threshold/count
-        storage.device.get_slip39_identifier(),
-        storage.device.get_slip39_iteration_exponent(),
+        identifier,
+        iteration_exponent,
         encrypted_master_secret,
     )[0]
 
@@ -144,12 +149,17 @@ async def backup_slip39_advanced(
         share_threshold = await layout.slip39_prompt_threshold(ctx, share_count, i)
         groups.append((share_threshold, share_count))
 
+    identifier = storage.device.get_slip39_identifier()
+    iteration_exponent = storage.device.get_slip39_iteration_exponent()
+    if identifier is None or iteration_exponent is None:
+        raise ValueError
+
     # generate the mnemonics
     mnemonics = slip39.split_ems(
         group_threshold=group_threshold,
         groups=groups,
-        identifier=storage.device.get_slip39_identifier(),
-        iteration_exponent=storage.device.get_slip39_iteration_exponent(),
+        identifier=identifier,
+        iteration_exponent=iteration_exponent,
         encrypted_master_secret=encrypted_master_secret,
     )
 
@@ -193,7 +203,7 @@ def _compute_secret_from_entropy(
 
 async def backup_seed(
     ctx: wire.Context, backup_type: BackupType, mnemonic_secret: bytes
-):
+) -> None:
     if backup_type == BackupType.Slip39_Basic:
         await backup_slip39_basic(ctx, mnemonic_secret)
     elif backup_type == BackupType.Slip39_Advanced:
