+++ core/src/trezor/ui/components/common/text.py
@@ -15,7 +15,7 @@ LINE_WIDTH = ui.WIDTH - TEXT_MARGIN_LEFT
 LINE_WIDTH_PAGINATED = LINE_WIDTH - PAGINATION_MARGIN_RIGHT
 
 if False:
-    from typing import Any, Union
+    from typing import Any, Sequence, Union
 
     TextContent = Union[str, int]
 
@@ -176,7 +176,7 @@ _WORKING_SPAN = Span()
 
 
 def render_text(
-    items: list[TextContent],
+    items: Sequence[TextContent],
     new_lines: bool,
     max_lines: int,
     font: int = ui.NORMAL,
